imrpoved git handler, removed cheatsheet

+++ shared/git_handler.py
@@ -1,5 +1,6 @@
 import requests
 from shared.ind_key import git
+import os
 
 def load_class_data_from_git(parent: str):
     # URL to the raw JSON file in the GitHub repository
@@ -13,17 +14,21 @@ def load_class_data_from_git(parent: str):
         "Authorization": f"token {token}"
     }
 
-    try:
-        # Send an HTTP GET request to the URL
-        response = requests.get(url, headers=headers)
-        
-        # Raise an exception for HTTP errors
-        response.raise_for_status()
-        
-        # Write the content of the response to a file
-        with open(local_filename, "w", encoding="utf-8") as file:
-            file.write(response.text)
-        
-        print(f"File downloaded successfully as {local_filename}")
-    except requests.exceptions.RequestException as e:
-        print(f"An error occurred: {e}")
+    if not os.path.isfile(local_filename):
+
+        try:
+            # Send an HTTP GET request to the URL
+            response = requests.get(url, headers=headers)
+            
+            # Raise an exception for HTTP errors
+            response.raise_for_status()
+            
+            # Write the content of the response to a file
+            with open(local_filename, "w", encoding="utf-8") as file:
+                file.write(response.text)
+            
+            print(f"File downloaded successfully as {local_filename}")
+        except requests.exceptions.RequestException as e:
+            print(f"An error occurred: {e}")
+    else:
+        print(f"{local_filename} already exists!")
